chore: update zsh setup to not require prompting

+++ scripts/setup.zsh.sh
@@ -1,24 +1,47 @@
 #!/usr/bin/env bash
 
-source ./scripts/util.sh
-
 function setupOhMyZshPlugins() {
-  git clone https://github.com/zsh-users/zsh-syntax-highlighting.git \
-    "${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting"
+  local plugin_dir="${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting"
+
+  if [ ! -d "$plugin_dir" ]; then
+    echo "Installing zsh-syntax-highlighting plugin..."
+    git clone https://github.com/zsh-users/zsh-syntax-highlighting.git "$plugin_dir"
+  else
+    echo "zsh-syntax-highlighting plugin already exists at $plugin_dir"
+  fi
 }
 
 function setupPurePrompt() {
-  mkdir -p "$HOME/.zsh"
-  rm -rf "$HOME/.zsh/pure"
-  git clone https://github.com/sindresorhus/pure.git "$HOME/.zsh/pure"
+  local pure_dir="$HOME/.zsh/pure"
+
+  if [ ! -d "$pure_dir" ]; then
+    echo "Installing pure prompt..."
+    mkdir -p "$HOME/.zsh"
+    rm -rf "$HOME/.zsh/pure"
+    git clone https://github.com/sindresorhus/pure.git "$HOME/.zsh/pure"
+  else
+    echo "pure prompt already exists at $pure_dir"
+  fi
 }
 
-runIfYes \
-  "pure-prompt" \
-  "Install pure-prompt?" \
-  setupPurePrompt;
+function setupZsh() {
+  if ! command -v zsh &> /dev/null; then
+    echo "Zsh not found. Installing..."
 
-runIfYes \
-  "oh-my-zsh plugins" \
-  "Install zsh-syntax-highlighting & other plugins?" \
-  setupOhMyZshPlugins;
+    if [[ $OSTYPE == 'darwin'* ]]; then
+      brew install zsh
+    else
+      sudo apt update
+      sudo apt install -y zsh
+    fi
+  fi
+
+  if [ ! -d "$HOME/.oh-my-zsh" ]; then
+    echo "Installing Oh My Zsh..."
+    sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)" "" --unattended
+  fi
+}
+
+setupZsh;
+setupOhMyZshPlugins;
+setupPurePrompt;
